Implement stats by date/time
Move stats functions and tests to separate files
